Now also parses installation directory

+++ infrastructure/suricata/docker-setup.sh
@@ -2,15 +2,13 @@
 
 function help() {
     echo -e "Usage:
-    $(basename $0) [ -l, --list ] [ -v, --verbose ] [ -h, --help ] \e[4mnetwork_interface"
+    $(basename $0) < -d, --directory > ${UNDERLINED}directory${RESET} [ -l, --list_interfaces ] [ -v, --verbose ] [ -h, --help ] ${UNDERLINED}network_interface${RESET}"
     
     exit 0
 }
 
 function list_interfaces() {
-    local network_interfaces=($(ls /sys/class/net))
-    
-    echo -e "${RED_BOLD}Your Network Interfaces are:${RESET}"
+	echo -e "${RED_BOLD}Your Network Interfaces are:${RESET}"
     
     for ((i = 1; i <= ${#network_interfaces[@]}; i++));
     do
@@ -21,28 +19,34 @@ function list_interfaces() {
 }
 
 RESET="\e[0m"
+UNDERLINED="\e[4m"
 RED_BOLD="\e[1;31m"
 GREEN="\e[32m"
 GREEN_BOLD="\e[1;32m"
 BG_RED_FG_WHITE="\e[41;1;37m"
 
-SHORT=l,v,h
-LONG=list,verbose,help
+SHORT=d:,l,v,h
+LONG=directory:,list_interfaces,verbose,help
 OPTS=$(getopt --name suricata-docker-setup --options $SHORT --longoptions $LONG -- "$@")
 
 eval set -- "$OPTS"
 
 VERBOSE="/dev/null"
+INSTALL_DIR=""
 
 while :
 do
     case "$1" in
-        -l | --list )
+		-d | --directory )
+			INSTALL_DIR="$2"
+			shift 2
+		;;
+        -l | --list_interfaces )
             list_interfaces
         ;;
         -v | --verbose )
             VERBOSE="/dev/stdout"
-            shift 2
+            shift
         ;;
         -h | --help )
             help
@@ -63,6 +67,14 @@ then
     echo -e "${RED_BOLD}You must pass the interface name to the script!${RESET}"
     help
 fi
+
+if [ -z "$INSTALL_DIR" ];
+then
+	echo -e "${RED_BOLD}You must pass the installation directory!${RESET}"
+	help
+fi
+
+INSTALL_DIR=$(realpath $INSTALL_DIR)
 
 network_interfaces=($(ls /sys/class/net))
 passed=false
@@ -89,28 +101,33 @@ then
     exit 0
 fi
 
-echo -e "${RED_BOLD}Creating /suricata... ${RESET}"
-mkdir /suricata &> "$VERBOSE"
-cd /suricata 
+echo -e "${RED_BOLD}Creating ${INSTALL_DIR}/... ${RESET}"
+mkdir -p "${INSTALL_DIR}/" &> "$VERBOSE"
+
+if [ ! -d "${INSTALL_DIR}/" ];
+then
+	echo -e "${RED_BOLD}Failed to create ${INSTALL_DIR}/...${RESET}"
+	exit 0
+fi
 
 echo -e "${RED_BOLD}Creating sub-directories for Docker Volumes: ${RESET}\n\
-	${GREEN}- /suricata/etc/ ${RESET}\n\
-	${GREEN}- /suricata/lib/ ${RESET}\n\
-	${GREEN}- /suricata/log/ ${RESET}"
-mkdir {etc,lib,log} &> "$VERBOSE"
+	${GREEN}- ${INSTALL_DIR}/etc/ ${RESET}\n\
+	${GREEN}- ${INSTALL_DIR}/lib/ ${RESET}\n\
+	${GREEN}- ${INSTALL_DIR}/log/ ${RESET}"
+mkdir {"${INSTALL_DIR}/etc/","${INSTALL_DIR}/lib/","${INSTALL_DIR}/log/"} &> "$VERBOSE"
 
 echo -e "${RED_BOLD}Running Suricata container (detatch)...${RESET}"
 docker run --rm -it -d --net=host \
 --cap-add=net_admin --cap-add=net_raw --cap-add=sys_nice \
--v $(pwd)/etc:/etc/suricata -v $(pwd)/lib:/var/lib/suricata -v $(pwd)/log:/var/log/suricata \
+-v "${INSTALL_DIR}/etc":/etc/suricata -v "${INSTALL_DIR}/lib":/var/lib/suricata -v "${INSTALL_DIR}/log":/var/log/suricata \
 --name=suricata jasonish/suricata:latest-amd64 -i "$1" &> "$VERBOSE"
 
 echo -e "${RED_BOLD}Fixing SIP, MQTT and RDP warnings.\n\
-Creating rules file /etc/suricata/ctf.rules${RESET}"
-touch /suricata/etc/ctf.rules &> "$VERBOSE"
+Creating rules file ${INSTALL_DIR}/etc/ctf.rules${RESET}"
+touch "${INSTALL_DIR}/etc/ctf.rules" &> "$VERBOSE"
 
-python3 -c 'import re
-FILE_PATH = "/suricata/etc/suricata.yaml"
+PYTHON_FIX='import re
+FILE_PATH = "'${INSTALL_DIR}'/etc/suricata.yaml"
 SIP_FIX = r"sip:\n      #enabled: no"
 MQTT_FIX = r"mqtt:\n      # enabled: no"
 RDP_FIX = r"rdp:\n      #enabled: yes"
@@ -122,7 +139,9 @@ with open(FILE_PATH, "r") as real_file:
     content = re.sub(RDP_FIX, "rdp:\n      enabled: no", content)
     content = re.sub(ADD_RULES, ADD_RULES + "\n  - /etc/suricata/ctf.rules", content)
 with open(FILE_PATH, "w") as modified_file:
-	modified_file.write(content)' &> "$VERBOSE"
+	modified_file.write(content)'
+	
+python3 -c "${PYTHON_FIX}" &> "$VERBOSE"
 
 echo -e "${RED_BOLD}Updating Suricata${RESET}"
 docker exec -it --user suricata suricata suricata-update -f &> "$VERBOSE"
